fix document scroller by refactoring css and code
